+++ repo-agent/pkg/llm/gemini.go
@@ -15,7 +15,6 @@
 package llm
 
 import (
-	"bytes"
 	"fmt"
 	"log"
 	"os"
@@ -86,31 +85,4 @@ func (g *Gemini) Run(agentPrompt string) ([]byte, error) {
 	}
 
 	return output, nil
-}
-
-// stripYAMLMarkers looks for ```yaml and ``` markers in the input byte slice.
-// If found, it strips these markers and returns the content between them.
-// If markers are not found, the original byte slice is returned.
-func stripYAMLMarkers(input []byte) ([]byte, error) {
-	startMarker := []byte("```yaml")
-	endMarker := []byte("```")
-
-	startIndex := bytes.Index(input, startMarker)
-	if startIndex == -1 {
-		return input, nil // Start marker not found
-	}
-
-	// Adjust startIndex to point after the start marker
-	startIndex += len(startMarker)
-
-	endIndex := bytes.Index(input[startIndex:], endMarker)
-	if endIndex == -1 {
-		return input, nil // End marker not found after start marker
-	}
-
-	// Adjust endIndex to be relative to the original input slice
-	endIndex += startIndex
-
-	// Extract the content between the markers, trimming any leading/trailing whitespace
-	return bytes.TrimSpace(input[startIndex:endIndex]), nil
 }
